// src/models/Event.ts
export interface Event {
  id: number;
  titre: string;
  description: string;
  lieu: string;
  date: string;
  disponible: boolean;
  nbrMax: number;
}
export interface MyEvent {
  id: number;
  titre: string;
  description: string;
  lieu: string;
  date: string;
  disponible: boolean;
  nbrMax: number;
}